US14042 - Android: Sign Up: Email: Did you mean TA31839 - code review
The following example was not matching in the web site:
- mobile: rob@x.c -> rob@me.com
- web: rob@x.c -> rob@x.com

The mobile app was picking up “x.c” as close to “me.com”, while the web site was not.

That was due to a subtle discrepancy between the Java and the JavaScript implementation of the Sift3 string distance algorithm. The Java implementation was rounding the calculation of the average string length while the JavaScript one was not.